% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{calc.stat}
\alias{calc.stat}
\title{Calculate the JLIM statistic}
\usage{
calc.stat(assoc1, assoc2, ld0, ld2, R2thr)
}
\arguments{
\item{assoc1}{data.frame The GWAS statistics as main_tr but with an additional
column, Z, containing the Z scores for each SNP}

\item{assoc2}{data.frame The eQTL statistics}

\item{ld0}{matrix The reference LD matrix for GWAS}

\item{ld2}{matrix The reference LD matrix for eQTLs (or can be same as ld0)}

\item{R2thr}{numeric The LD window threshold for calculating the JLIM statistic.
SNP significance is compared from within to outside this window}
}
\value{
The numeric JLIM statistic for the test
}
\description{
Calculate the JLIM statistic
}
