% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{calcMAF.refPanels}
\alias{calcMAF.refPanels}
\title{Filter rare variants from the reference haplotypes}
\usage{
calcMAF.refPanels(refgt, refgt.all.sel)
}
\arguments{
\item{refgt}{data.frame Reference haplotypes with columns of CHROM, POS, REF, ALT, GT, 
followed by the base of each each individual. Needs to use same reference genome as your GWAS and eQTL
genotype data. (default=NULL)}

\item{refgt.all.sel}{logical Indicates which rows of refgt to select}
}
\value{
a filtered refgt object
}
\description{
Filter rare variants from the reference haplotypes
}
