% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{find.panel}
\alias{find.panel}
\title{Find the proper ref-panel from the list of the files in the specified path}
\usage{
find.panel(refLD.dir, start.bp, end.bp, CHR)
}
\arguments{
\item{refLD.dir}{character The directory containing the reference haplotype files}

\item{start.bp}{integer The starting BP of the locus to test}

\item{end.bp}{integer The ending BP of the locus to test}

\item{CHR}{integer The chr of the locus to test}
}
\value{
The character name of the file to load
}
\description{
Find the proper ref-panel from the list of the files in the specified path
}
