% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{get_acat_cells}
\alias{get_acat_cells}
\title{Determine which cells contribute to the overall significance}
\usage{
get_acat_cells(pvals, null_dist, pv_thresh = 0.1, initial_step_size = 1000)
}
\arguments{
\item{pvals}{numeric The per cell pvalues output from jlim_main().}

\item{null_dist}{list The output from get_null_dist()}

\item{pv_thresh}{numeric The cauchy p-value at which to stop including cells}

\item{initial_step_size}{numeric The step size to use initially. Helpful for
it to be roughly 1/10 the number of cells.}
}
\value{
A character vector with the names of the significantly contributing cells
}
\description{
Determine which cells contribute to the overall significance
}
