% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{get_null_dist}
\alias{get_null_dist}
\title{Generate a null distribution of JLIM statistics}
\usage{
get_null_dist(
  jlim_vars,
  sectr.sample.size,
  nperm,
  n.cores,
  r2res = 0.8,
  progress = TRUE
)
}
\arguments{
\item{jlim_vars}{list Preprocessed data objects from prep_jlim()}

\item{sectr.sample.size}{numeric The number of individuals for eQTL mapping}

\item{nperm}{numeric The size of the null distribution}

\item{n.cores}{numeric The number of cores to use}

\item{r2res}{numeric the LD window threshold for calculating the JLIM statistic.
SNP significance is compared from within to outside this window. (default=0.8)}

\item{progress}{logical To show a progress bar (default=TRUE)}
}
\value{
a list containing the null distribution vector in the first element and the 
r2res parameter in the second
}
\description{
Generate a null distribution of JLIM statistics
}
