% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_mapping.R
\name{get_per_cell_pv_covar}
\alias{get_per_cell_pv_covar}
\title{Compute total effects, errors, and p-values per cell for one SNP}
\usage{
get_per_cell_pv_covar(data_in, coef_mat, n_PCs, vcov_mat, get_beta = FALSE)
}
\arguments{
\item{data_in}{data.frame A data object with cells as rows. Expression values
are in the first column with PCs in the next columns, followed by donors and
other covariate data.}

\item{coef_mat}{data.frame The first list element output from lme_helper(), containing
the mixed model term coefficients, standard errors, and p-values.}

\item{n_PCs}{numeric The number of PCs and PC interaction terms used.}

\item{vcov_mat}{matrix The second list element output from lme_helper(), containing
the variance and covariances between each model term.}

\item{get_beta}{logical Whether to return beta values in addition to p-values.
Only returns p-values if FALSE. (default=FALSE)}
}
\value{
a numeric vector containing the p-values for all cells
}
\description{
Compute total effects, errors, and p-values per cell for one SNP
}
