% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{get_permmat}
\alias{get_permmat}
\title{Generate a matrix of null eQTL z-scores (nperm x snps)}
\usage{
get_permmat(refgt_num, sectr.sample.size, nperm, n.cores, progress)
}
\arguments{
\item{refgt_num}{data.frame The reference panel with ref/alt as 1s and 2s}

\item{sectr.sample.size}{numeric The number of eQTL individuals}

\item{nperm}{numeric The size of the null distribution}

\item{n.cores}{numeric The number of cores to use}

\item{progress}{logical Whether to show progress}
}
\value{
a matrix of null eQTL Z-scores
}
\description{
Generate a matrix of null eQTL z-scores (nperm x snps)
}
