% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{jlim.test.numeric}
\alias{jlim.test.numeric}
\title{Run a single JLIM test}
\usage{
jlim.test.numeric(
  jlim_vars,
  null_dist,
  sec_tr,
  sectr.sample.size,
  min.SNPs.count
)
}
\arguments{
\item{jlim_vars}{list Preprocessed data objects from prep_jlim()}

\item{null_dist}{list The output from get_null_dist()}

\item{sec_tr}{data.frame The eQTL results for one cell, with SNPs as rows and 
CHR, BP, and P as columns}

\item{sectr.sample.size}{numeric The number of eQTL individuals}

\item{min.SNPs.count}{numeric The minimum number of SNPs needed to run}
}
\value{
a jlim object containing the test results
}
\description{
Run a single JLIM test
}
