% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{perm.test}
\alias{perm.test}
\title{Generate the null statistics for null simulated eQTL data}
\usage{
perm.test(assoc1, permmat, ld0, ld2, r2res, n.cores, progress = progress)
}
\arguments{
\item{assoc1}{data.frame The GWAS statistics as main_tr but with an additional
column, Z, containing the Z scores for each SNP}

\item{permmat}{matrix Z-scores for the null eQTL data}

\item{ld0}{matrix The LD correlation matrix}

\item{ld2}{matrix The LD correlation matrix (can be the same as ld0 or different if
specifically for the eQTL individuals)}

\item{r2res}{numeric The LD window threshold for calculating the JLIM statistic.
SNP significance is compared from within to outside this window}

\item{n.cores}{numeric The number of cores to use in parallelization}

\item{progress}{logical Whether to show progress}
}
\value{
A vector of the null JLIM statistics
}
\description{
Generate the null statistics for null simulated eQTL data
}
