% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{perm_test_run}
\alias{perm_test_run}
\title{Function to run the permutation test Rcpp fn}
\usage{
perm_test_run(
  assoc1,
  permmat = permmat,
  ld0 = ld_cormat,
  ld2 = ld_cormat,
  R2thr = 0.8,
  lambda_t = Inf
)
}
\arguments{
\item{assoc1}{}

\item{permmat}{}

\item{ld0}{}

\item{ld2}{}

\item{R2thr}{}

\item{lambda_t}{}
}
\description{
Function to run the permutation test Rcpp fn
}
