% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_mapping.R
\name{pme_helper}
\alias{pme_helper}
\title{Fit a mixed model for one SNP and extract coefficients}
\usage{
pme_helper(data_in, geno_mat, snp_name, n_PCs, julia)
}
\arguments{
\item{data_in}{data.frame A data object with cells as rows. Expression values
are in the first column with PCs in the next columns, followed by donors and
other covariate data.}

\item{geno_mat}{data.frame Donor genotype data (SNPs x donors)}

\item{snp_name}{character The name of the SNP to test. Should match the name
of a row from geno_mat.}

\item{n_PCs}{numeric The number of PCs and PC-interaction terms to use}

\item{julia}{environment The output of julia_setup()}
}
\value{
a list with the coefficients matrix in the first element and the
variance-covariance matrix in the second element
}
\description{
Fit a mixed model for one SNP and extract coefficients
}
