% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{prep_jlim}
\alias{prep_jlim}
\title{Prepares raw data objects for running scJLIM}
\usage{
prep_jlim(main_tr, refLD.dir = NULL, refLD.mat = NULL, min.MAF = 0.05)
}
\arguments{
\item{main_tr}{data.frame The GWAS summary statistics with columns of CHR, 
BP, and P}

\item{refLD.dir}{character Directory location of reference haplotype directory
when different segments are in separate files. For example, you can download 
https://www.dropbox.com/s/1uaw5zlpwhz3ipz/refld.1kg.nfe.b37.tar.gz?dl=0
for hg19 1000GP Non-Finnish Europeans (NFE) and set this parameter to the directory
/refld.1kg.nfe.b37. This file and others can be also be found here 
https://github.com/cotsapaslab/jlim. (default=NULL)}

\item{refLD.mat}{data.frame Required if refLD.dir is not supplied. Reference 
haplotypes with columns of CHROM, POS, REF, ALT, GT, followed by the base of 
each each individual. Needs to use same reference genome as your GWAS and eQTL
genotype data. (default=NULL)}

\item{min.MAF}{numeric A minimum threshold to select SNPs for analysis (default=.05)}
}
\value{
a list of data objects for use in downstream JLIM functions
}
\description{
Prepares raw data objects for running scJLIM
}
