% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{process.refPanels}
\alias{process.refPanels}
\title{Convert matrix of genotypes into matrix of 1s and 2s for ref/alt. Also 
get the LD matrix of pearson correlations.}
\usage{
process.refPanels(refgt.org)
}
\arguments{
\item{refgt.org}{data.frame MAF thresholded reference haplotypes with the 
same column variables as before}
}
\value{
list with the LD matrix and modified haplotype data
}
\description{
Convert matrix of genotypes into matrix of 1s and 2s for ref/alt. Also 
get the LD matrix of pearson correlations.
}
