% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{PtoZ}
\alias{PtoZ}
\title{Converts p-values to Z-scores}
\usage{
PtoZ(pv)
}
\arguments{
\item{pv}{numeric A p-value}
}
\value{
A z-score corresponding to the p-value
}
\description{
Converts p-values to Z-scores
}
