% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{SNPselction.numeric}
\alias{SNPselction.numeric}
\title{Helper function for running JLIM - adapted from main JLIM package}
\usage{
SNPselction.numeric(
  assoc1,
  assoc2,
  ld1,
  ld2,
  ld0.maf,
  r2res,
  sectr.sample.size,
  min.SNPs.count,
  indexSNP,
  NULLDIST
)
}
\arguments{
\item{assoc1}{data.frame The GWAS summary statistics with Z column added}

\item{assoc2}{data.frame The eQTL summary statistics for a cell}

\item{ld1}{matrix The reference LD matrix for GWAS}

\item{ld2}{matrix The reference LD matrix for eQTL (can just use same as ld1)}

\item{ld0.maf}{numeric MAFs for the SNPs in LD matrix, with BP as names}

\item{r2res}{numeric The LD window threshold for calculating the JLIM statistic.
SNP significance is compared from within to outside this window}

\item{sectr.sample.size}{numeric The number of eQTL individuals. Should be the
same as ncol(geno_mat) as used in get_eQTL_res().}

\item{min.SNPs.count}{numeric The minimum number of SNPs needed to run}

\item{indexSNP}{integer The BP of the lead GWAS SNP}

\item{NULLDIST}{numeric The null distribution of JLIM statistics}
}
\value{
a jlim object containing the results of the test
}
\description{
Helper function for running JLIM - adapted from main JLIM package
}
