% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{stdGT2}
\alias{stdGT2}
\title{Standardize 0s and 1s matrix of genotypes}
\usage{
stdGT2(gt)
}
\arguments{
\item{gt}{matrix Simulated genotypes from reference haplotypes}
}
\value{
a standardized reference genotypes matrix
}
\description{
Standardize 0s and 1s matrix of genotypes
}
