% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{toGT}
\alias{toGT}
\title{Helper function that creates genotypes from haplotypes by summing number of
alternate alleles from two random haplotypes}
\usage{
toGT(gt)
}
\arguments{
\item{gt}{matrix Contains reference haplotypes as 1s and 2s}
}
\value{
a matrix of simulated genotypes from haplotypes
}
\description{
Helper function that creates genotypes from haplotypes by summing number of
alternate alleles from two random haplotypes
}
