% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_mapping.R
\name{get_eQTL_res}
\alias{get_eQTL_res}
\title{Main function to run eQTL mapping}
\usage{
get_eQTL_res(
  gene_test,
  norm_counts,
  cell_meta,
  cell_pcs,
  n_PCs,
  geno_mat,
  main_tr,
  julia_dir,
  geno_pcs = NULL,
  n.cores = 4,
  progress = TRUE
)
}
\arguments{
\item{gene_test}{character The gene to test for cis-eQTLs with}

\item{norm_counts}{matrix Normalized UMI counts matrix (genes x cells)}

\item{cell_meta}{data.frame Cell-level metadata. Columns must include 'donors', as
the model includes donor random effects. Rownames must be cell names. Optionally include a
'batch' column to also incorporate batch random effects. Include all other covariates to be
used in the model. We suggest pre-scaling and log transforming variables with long tails such as library sizes.}

\item{cell_pcs}{matrix Cell by PC matrix, used for genotype-interaction testing. Columns
should be labeled PC1, PC2, etc. Rownames should be cell names.}

\item{n_PCs}{numeric The number of cell_pcs to use for interaction testing}

\item{geno_mat}{matrix Imputed genotype matrix (SNPs by donors), with values representing
the number of alternate alleles per donor (0, 1, or 2)}

\item{main_tr}{data.frame The GWAS locus data with columns as 'CHR','BP','P', corresponding
to chromosome, position, and p-value}

\item{julia_dir}{character The directory of your Julia installation}

\item{geno_pcs}{matrix Optional donor by PC matrix to include as covariates in the model.
Columns should be labeled as geno_PC1, geno_PC2, etc. (default=NULL)}

\item{n.cores}{numeric Number of cores to use (default=4)}

\item{progress}{logical Whether to show a progress bar (default=TRUE)}
}
\value{
A list with one element per cell. Each element contains a vector with the p-values
for each tested SNP for the given cell.
}
\description{
Main function to run eQTL mapping
}
