% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{jlim_main}
\alias{jlim_main}
\title{The main user-facing function for running JLIM colocalization test per cell}
\usage{
jlim_main(
  snp_res_mat,
  jlim_vars,
  null_dist,
  sectr.sample.size,
  min.SNPs.count = 15,
  n.cores = 20
)
}
\arguments{
\item{snp_res_mat}{list The output from get_eQTL_res(). Contains one element
per cell with the p-value for each SNP in that cell.}

\item{jlim_vars}{list Preprocessed data objects from prep_jlim()}

\item{null_dist}{list The output from get_null_dist()}

\item{sectr.sample.size}{numeric The number of eQTL individuals. Should be the
same as ncol(geno_mat) as used in get_eQTL_res().}

\item{min.SNPs.count}{numeric The minimum number of SNPs needed to run a test
(default=15)}

\item{n.cores}{numeric The number of cores to use when running in parallel (default=20)}
}
\value{
a list with the cauchy global p-value in the first element and the per
cell p-values in the second element.
}
\description{
The main user-facing function for running JLIM colocalization test per cell
}
