% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLIM_functions.R
\name{loadRefLD}
\alias{loadRefLD}
\title{Load the reference panel and filters out low MAF variants}
\usage{
loadRefLD(refgt, start.bp, end.bp, min.MAF)
}
\arguments{
\item{refgt}{data.frame Reference haplotypes with columns of CHROM, POS, REF, ALT, GT, 
followed by the base of each each individual. Needs to use same reference genome as your GWAS and eQTL
genotype data. (default=NULL)}

\item{start.bp}{numeric The starting BP of the locus to test}

\item{end.bp}{numeric The ending BP of the locus to test}

\item{min.MAF}{numeric The minimum MAF threshold for selecting SNPs}
}
\value{
list containing the ref LD matrix and MAFs of each SNP
}
\description{
Load the reference panel and filters out low MAF variants
}
